LayoutTests/imported/w3c/web-platform-tests/html-aam/roles-contextual-expected.txt

Tests contextual computed role mappings defined in HTML-AAM, where the returned computed role is expected to change based on the context. Most test names correspond to a unique ID defined in the spec.

These should remain in alphabetical order.

x x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x
x

x
x
x
x
x
x
x
x
x
x
x
labelledby

PASS el-a
PASS el-aside
PASS el-aside-in-main
PASS el-aside-in-article-in-main-with-name
PASS el-aside-in-article-with-name
PASS el-aside-in-aside-with-name
PASS el-aside-in-nav-with-name
PASS el-aside-in-nav-with-role
PASS el-aside-in-section-with-name
PASS el-aside-in-section-aria-labelledby
PASS el-aside-in-section-title
PASS el-footer-ancestorbody
PASS el-header-ancestorbody
FAIL el-img-no-name assert_equals: <img data-testname="el-img-no-name" data-expectedrole="image" class="ex" src="data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw=="> expected "image" but got "none"
FAIL el-img-empty-alt-aria-label assert_equals: <img data-testname="el-img-empty-alt-aria-label" data-expectedrole="image" class="ex" src="data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==" alt="" aria-label="x"> expected "image" but got "none"
FAIL el-img-empty-alt-aria-labelledby assert_equals: <img data-testname="el-img-empty-alt-aria-labelledby" data-expectedrole="image" class="ex" src="data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==" alt="" aria-labelledby="labelledby"> expected "image" but got "none"
PASS el-section
PASS el-section-aria-labelledby
PASS el-section-title
PASS el-a-no-href
PASS el-aside-in-article-in-main
PASS el-aside-in-article
PASS el-aside-in-aside
PASS el-aside-in-nav
PASS el-aside-in-section
PASS el-aside-in-section-aria-label-empty
PASS el-aside-in-section-aria-label-whitespace
FAIL el-aside-in-section-aria-labelledby-non-existing assert_false: Computed Role: "complementary" does not match any of the acceptable role strings in ["generic", "", "none"]: <aside data-testname="el-aside-in-section-aria-labelledby-non-existing" class="ex-generic" aria-labelledby="non-existing">x</aside> expected false got true
FAIL el-aside-in-section-aria-labelledby-empty assert_false: Computed Role: "complementary" does not match any of the acceptable role strings in ["generic", "", "none"]: <aside data-testname="el-aside-in-section-aria-labelledby-empty" class="ex-generic" aria-labelledby="empty">x</aside> expected false got true
FAIL el-aside-in-section-aria-labelledby-whitespace assert_false: Computed Role: "complementary" does not match any of the acceptable role strings in ["generic", "", "none"]: <aside data-testname="el-aside-in-section-aria-labelledby-whitespace" class="ex-generic" aria-labelledby="space">x</aside> expected false got true
PASS el-aside-in-section-title-empty
FAIL el-aside-in-section-title-whitespace assert_false: Computed Role: "complementary" does not match any of the acceptable role strings in ["generic", "", "none"]: <aside data-testname="el-aside-in-section-title-whitespace" class="ex-generic" title=" ">x</aside> expected false got true
PASS el-img-empty-alt-aria-label-empty
PASS el-img-empty-alt-aria-label-whitespace
PASS el-img-empty-alt-aria-labelledby-non-existing
PASS el-img-empty-alt-aria-labelledby-empty
PASS el-img-empty-alt-aria-labelledby-whitespace
PASS el-img-empty-alt-title
PASS el-img-empty-alt-title-empty
PASS el-img-empty-alt-title-whitespace
PASS el-section-no-name
PASS el-section-aria-label-empty
PASS el-section-aria-label-whitespace
FAIL el-section-aria-labelledby-non-existing assert_false: Computed Role: "region" does not match any of the acceptable role strings in ["generic", "", "none"]: <section data-testname="el-section-aria-labelledby-non-existing" class="ex-generic" aria-labelledby="non-existing">x</section> expected false got true
FAIL el-section-aria-labelledby-empty assert_false: Computed Role: "region" does not match any of the acceptable role strings in ["generic", "", "none"]: <section data-testname="el-section-aria-labelledby-empty" class="ex-generic" aria-labelledby="empty">x</section> expected false got true
FAIL el-section-aria-labelledby-whitespace assert_false: Computed Role: "region" does not match any of the acceptable role strings in ["generic", "", "none"]: <section data-testname="el-section-aria-labelledby-whitespace" class="ex-generic" aria-labelledby="space">x</section> expected false got true
PASS el-section-title-empty
FAIL el-section-title-whitespace assert_false: Computed Role: "region" does not match any of the acceptable role strings in ["generic", "", "none"]: <section data-testname="el-section-title-whitespace" class="ex-generic" title=" ">x</section> expected false got true
